Split some files
